requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
boto3==1.24.28
botocore==1.27.59
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
evaluate==0.4.0
filelock==3.9.0
fonttools==4.25.0
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.2
munkres==1.1.4
networkx==3.1
nltk==3.7
numexpr==2.8.4
numpy==1.23.5
omegaconf==2.3.0
openai==0.27.2
openpyxl==3.0.9
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230321
pathtools==0.1.2
Pillow==9.4.0
pip==23.0.1
platformdirs==3.2.0
plotly==5.14.0
pooch==1.7.0
protobuf==3.20.3
psutil==5.9.0
pycparser==2.21
Pygments==2.15.0
pynvml==11.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
regex==2022.7.9
requests==2.28.2
rich==13.3.3
rouge_score==0.1.2
s3transfer==0.6.0
sacremoses==0.0.43
scalene==1.5.20
scikit-learn==1.2.2
scipy==1.10.0
sentry-sdk==1.18.0
setproctitle==1.2.2
setuptools==65.6.3
six==1.16.0
smmap==3.0.5
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
types-pytz==2023.3.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wheel==0.38.4
yarl==1.8.1
zipp==3.11.0
